some extra math functions

+++ extra_math.h
@@ -6,7 +6,7 @@
 /*   By: isemin <isemin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/29 09:42:08 by isemin            #+#    #+#             */
-/*   Updated: 2025/01/29 09:52:03 by isemin           ###   ########.fr       */
+/*   Updated: 2025/01/29 09:52:49 by isemin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 # include "types_numeric.h"
 
-float	distance(t_DoublePair a, t_DoublePair b);
-float	ceil_map(float x);
-float	floor_map(float x);
+double	distance(t_DoublePair a, t_DoublePair b);
+double	ceil_map(double x);
+double	floor_map(double x);
 
 #endif
